Create wallet component
