+++ package.json
@@ -7,8 +7,8 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
-    "firebase": "^9.8.3",
     "axios": "^0.27.2",
+    "firebase": "^9.8.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
